Less25

+++ src/homeworks/homework_24/Claculator.java
@@ -30,8 +30,9 @@ public class Claculator {
         return sum;
     }
 
-    public static int dividing(int num1, int num2) {
-        int sum = num1 / num2;
+    public static double dividing(double num1, double num2) {
+//        if (num2 == 0) return Integer.MAX_VALUE; // если делить тип данных int, чтобы программа не уходила в ошибку.
+        double sum = num1 / num2;
         System.out.print("Результат деления = ");
         return sum;
     }
